requirements.txt
feedparser
google-genai
streamlit
pandas
python-dotenv

pymysql
